ensure space checks work

+++ attack_simulator/mal/observation.py
@@ -25,18 +25,18 @@ class Observation:
     def from_rust(cls, obs):
         state, assets, asset_ids, names = zip(*obs.nodes)
         return Observation(
-            ids_observation=np.array(obs.ids_observation),
-            attack_surface=np.array(obs.attack_surface),
-            defense_surface=np.array(obs.defense_surface),
-            state=np.array(state),
-            assets=np.array(assets),
-            asset_ids=np.array(asset_ids),
-            names=np.array(names),
-            ttc_remaining=np.array(obs.ttc_remaining),
-            defender_action_mask=np.array(obs.defender_action_mask),
-            attacker_action_mask=np.array(obs.attacker_action_mask),
-            edges=np.array(obs.edges),
-            flags=np.array(obs.flags),
+            ids_observation=np.array(obs.ids_observation, dtype=np.int8),
+            attack_surface=np.array(obs.attack_surface, dtype=np.int8),
+            defense_surface=np.array(obs.defense_surface, dtype=np.int8),
+            state=np.array(state, dtype=np.int8),
+            assets=np.array(assets, dtype=np.uint64),
+            asset_ids=np.array(asset_ids, dtype=np.uint64),
+            names=np.array(names, dtype=np.uint64),
+            ttc_remaining=np.array(obs.ttc_remaining, dtype=np.uint64),
+            defender_action_mask=np.array(obs.defender_action_mask, dtype=np.int8),
+            attacker_action_mask=np.array(obs.attacker_action_mask, dtype=np.int8),
+            edges=np.array(obs.edges, dtype=np.uint64),
+            flags=np.array(obs.flags, dtype=np.uint64),
         )
